Move compiler and its utilities to a service dir
